Change a little bit the transform implementation & fix wrong index size

+++ source/shogle/render/shaders/generic3d.hpp
@@ -32,13 +32,13 @@ public:
 
   inline generic3d& bind_diffuse(const gl::texture& tex) {
     tex.bind(_diffuse_sampler);
-    set_uniform(_diffuse_unif, _diffuse_sampler);
+    set_uniform(_diffuse_unif, (int)_diffuse_sampler);
     return *this;
   }
 
   inline generic3d& bind_specular(const gl::texture& tex) {
     tex.bind(_specular_sampler);
-    set_uniform(_specular_unif, _specular_sampler);
+    set_uniform(_specular_unif, (int)_specular_sampler);
     return *this;
   }
